+++ client/coral-framework/index.js
@@ -4,7 +4,6 @@ import * as itemActions from './actions/items';
 import I18n from './modules/i18n/i18n';
 import * as notificationActions from './actions/notification';
 import * as authActions from './actions/auth';
-import * as configActions from './actions/config';
 import * as assetActions from './actions/asset';
 
 export {
@@ -14,6 +13,5 @@ export {
   I18n,
   notificationActions,
   authActions,
-  configActions,
   assetActions
 };
