+++ client/coral-framework/reducers/index.js
@@ -2,7 +2,6 @@ import auth from './auth';
 import user from './user';
 import asset from './asset';
 import items from './items';
-import config from './config';
 import notification from './notification';
 
 export default {
@@ -10,6 +9,5 @@ export default {
   user,
   asset,
   items,
-  config,
   notification
 };
